Crate onnxruntime_sys

Structs

OrtAllocator

\brief Memory allocation interface

OrtApi

\brief The C API

OrtApiBase

\brief The helper interface to get the right version of OrtApi

OrtArenaCfg

OrtCANNProviderOptions

OrtCUDAProviderOptions

\brief CUDA Provider Options

OrtCUDAProviderOptionsV2

OrtCustomHandleType

OrtCustomOp

OrtCustomOpDomain

OrtDnnlProviderOptions

OrtEnv

OrtIoBinding

OrtKernelContext

OrtKernelInfo

OrtMIGraphXProviderOptions

\brief MIGraphX Provider Options

OrtMapTypeInfo

OrtMemoryInfo

OrtModelMetadata

OrtOp

OrtOpAttr

OrtOpenVINOProviderOptions

\brief OpenVINO Provider Options

OrtPrepackedWeightsContainer

OrtROCMProviderOptions

\brief ROCM Provider Options

OrtRunOptions

OrtSequenceTypeInfo

OrtSession

OrtSessionOptions

OrtStatus

OrtTensorRTProviderOptions

\brief TensorRT Provider Options

OrtTensorRTProviderOptionsV2

OrtTensorTypeAndShapeInfo

OrtThreadPoolParams

OrtThreadingOptions

OrtTrainingApi

OrtTypeInfo

OrtValue

onnxruntime

Enums

ExecutionMode

GraphOptimizationLevel

\brief Graph optimization level

ONNXTensorElementDataType

Copied from TensorProto::DataType Currently, Ort doesn’t support complex64, complex128

ONNXType

OrtAllocatorType

OrtCudnnConvAlgoSearch

\brief Algorithm to use for cuDNN Convolution Op

OrtCustomOpInputOutputCharacteristic

OrtErrorCode

OrtLanguageProjection

\brief Language projection identifiers /see OrtApi::SetLanguageProjection

OrtLoggingLevel

\brief Logging severity levels

OrtMemType

\brief Memory types for allocated memory, execution provider specific types should be extended in each provider.

OrtMemoryInfoDeviceType

\brief This mimics OrtDevice type constants so they can be returned in the API

OrtOpAttrType

OrtSparseFormat

OrtSparseIndicesFormat

Constants

ORT_API_VERSION

Functions

library_filename

Converts a library name to a filename generally appropriate for use on the system.

Type Definitions

OnnxEnumInt

OrtCustomCreateThreadFn

\brief Ort custom thread creation function

OrtCustomJoinThreadFn

\brief Custom thread join function

OrtCustomThreadHandle

OrtLoggingFunction

OrtStatusPtr

OrtThreadWorkerFn

\brief Thread work loop function